requirements.txt
# AIMO Analysis Engine - Python Dependencies
# 
# Python 3.11+ required
# Install: pip install -r requirements.txt
# For development: pip install -r requirements.txt -r requirements-dev.txt

# =============================================================================
# Core Dependencies
# =============================================================================

# Database
duckdb==1.1.3

# Data Processing
polars==1.20.0
pyarrow==18.1.0

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# Date/Time
python-dateutil==2.9.0.post0

# Domain Parsing (eTLD+1 / Public Suffix List)
tldextract==5.1.2

# JSON Schema Validation
jsonschema==4.23.0

# =============================================================================
# LLM Integration
# =============================================================================

# HTTP Client (async support)
httpx==0.28.1

# HTTP Client (sync, for LLM client)
# Note: urllib3 1.x for LibreSSL compatibility (urllib3 2.x requires OpenSSL 1.1.1+)
requests==2.32.3
urllib3>=1.26.0,<2.0.0

# Retry with Backoff
tenacity==9.0.0

# Google GenAI SDK (for Gemini)
# Note: 1.48.0+ requires Python 3.10+, using 1.47.0 for Python 3.9 compatibility
google-genai>=1.47.0,<1.48.0

# =============================================================================
# File Locking
# =============================================================================

filelock==3.16.1

# =============================================================================
# Excel Generation
# =============================================================================

# XlsxWriter for constant-memory Excel generation
xlsxwriter==3.2.0

# openpyxl for reading/template-based operations
openpyxl==3.1.5

# =============================================================================
# String Matching (for fuzzy service name matching)
# =============================================================================

rapidfuzz==3.10.1

# =============================================================================
# Optional: System Monitoring
# =============================================================================

psutil==6.1.1

# =============================================================================
# Development Dependencies (separate file recommended)
# =============================================================================

# pytest>=8.0.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.23.0
# ruff>=0.4.0
# mypy>=1.10.0
